refactor: extract ws conn handling as custom hook
